Move spare parts enums to a separate folder
