+++ PocketForzaHorizonCommunity.Back/PocketForzaHorizonCommunity.Back.Database/Enums/SpareParts/TiresWidthType.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace PocketForzaHorizonCommunity.Back.Database.Enums;
+namespace PocketForzaHorizonCommunity.Back.Database.Enums.SpareParts;
 
 [JsonConverter(typeof(StringEnumConverter))]
 public enum TiresWidthType
